abstract on ComputeStepFactory
