+++ algo/src/main/java/org/neo4j/graphalgo/impl/pagerank/WeightedComputeStepFactory.java
@@ -4,9 +4,9 @@ import org.neo4j.graphalgo.api.Degrees;
 import org.neo4j.graphalgo.api.RelationshipIterator;
 import org.neo4j.graphalgo.api.RelationshipWeights;
 
-public class NonWeightedComputeStepFactory implements  ComputeStepFactory {
+public class WeightedComputeStepFactory implements  ComputeStepFactory {
     public ComputeStep createComputeStep(double dampingFactor, int[] sourceNodeIds, RelationshipIterator relationshipIterator, Degrees degrees, RelationshipWeights relationshipWeights, int partitionCount, int start) {
-        return new NonWeightedComputeStep(
+        return new WeightedComputeStep(
                 dampingFactor,
                 sourceNodeIds,
                 relationshipIterator,
